versions/v4/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework-simplejwt==5.3.0
python-multipart==0.0.6
email-validator==2.1.0
bcrypt==4.1.2